+++ backend/config/database.py
@@ -2,7 +2,7 @@ from sqlalchemy import create_engine, MetaData
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
-SQLALCHEMY_DATABASE_URL = ""
+SQLALCHEMY_DATABASE_URL = "postgresql://postgres:71440ed75e0cde0b993d560f1f54d8b1@00-default.gabrielrosa.dev:10323/benchml"
 
 engine = create_engine(
     SQLALCHEMY_DATABASE_URL
