+++ backend/main.py
@@ -1,9 +1,9 @@
 from fastapi import FastAPI, Depends, HTTPException
 from sqlalchemy.orm import Session
 from typing import List
-from controller import crud
-from database import schemas
-from config.database import SessionLocal
+from backend.controller import crud
+from backend.database import schemas
+from backend.config.database import SessionLocal
 
 app = FastAPI()
